etrms/backend/requirements.txt
# Enhanced Trading Risk Management System Dependencies
# Python 3.9+ required

# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
sqlalchemy==2.0.23
python-dotenv==1.0.0
python-multipart==0.0.6

# Authentication dependencies
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
psycopg2-binary==2.9.9
alembic==1.12.1

# WebSockets
websockets==12.0

# Redis for caching and pub/sub
redis==5.0.1

# HTTP client
httpx==0.25.1
requests==2.31.0

# Data processing
pandas==2.1.2
numpy==1.26.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Logging
loguru==0.7.2

# Development tools
black==23.10.1
isort==5.12.0
flake8==6.1.0

# Backtesting dependencies
tqdm==4.66.1                 # Progress bars for data downloading
matplotlib==3.8.0            # Plotting for backtesting reports
seaborn==0.13.0              # Enhanced visualizations for backtesting reports
jinja2==3.1.2                # Templates for HTML reports
pytz==2023.3                 # Timezone handling

# Additional dependencies
scipy==1.10.1                # Scientific computing
scikit-learn==1.2.1          # Machine learning
statsmodels==0.13.5          # Statistical models
python-binance==1.0.16       # Binance API client
python-jose==3.3.0           # JWT token handling 